feat: implement CalloutItem component for improved FAQ navigation
