+++ src/app/faqs/page.tsx
@@ -1,34 +1,18 @@
 "use client"
 
-import { useRouter } from "next/navigation"
-import { FiArrowUpRight } from "react-icons/fi"
-import { Button } from "@/components/ui/button"
-import { Item, ItemActions, ItemContent, ItemInner, ItemTitle } from "@/components/ui/item"
+import CalloutItem from "@/components/callout-item"
 
 export default function FAQsPage() {
-  const router = useRouter()
-
   return (
     <main className="w-full">
       <div className="flex flex-col items-center py-12">
-        <Item className="flex w-full max-w-4xl flex-col">
-          <ItemInner>
-            <ItemContent>
-              <ItemTitle>Non trovi ciò che stai cercando?</ItemTitle>
-            </ItemContent>
-            <ItemActions>
-              <Button
-                variant="primary"
-                size="lg"
-                className="gap-8 pr-6 pl-10 has-[>svg]:pr-6 has-[>svg]:pl-10"
-                onClick={() => router.push("/guides")}
-              >
-                Esplora le Guide
-                <FiArrowUpRight aria-hidden="true" />
-              </Button>
-            </ItemActions>
-          </ItemInner>
-        </Item>
+        <CalloutItem
+          title="Non trovi ciò che stai cercando?"
+          href="/guides"
+          buttonText="Esplora le Guide"
+          className="max-w-4xl"
+        />
+
       </div>
     </main>
   )
