+++ src/components/callout-item/types.ts
@@ -1,0 +1,6 @@
+export type CalloutItemProps = {
+    title: string
+    href: string
+    buttonText: string
+    className?: string
+}
